ressurect_knights now performs basic checks to determine whether the input file is of the correct format.

+++ knight.py
@@ -438,7 +438,6 @@ def get_knights(lines):
             knight.append(line)
     return kns
 
-# TODO allow for names with multiple words in them
 # TODO Add error handling for files with incorrect formatting
 # Takes the 4 string lists from which are returned as elements of a list from get_knights
 # and extracts the info necessary to rebuild the description for further use.
@@ -450,9 +449,21 @@ def retrieve_details(kn):
     castle = ' '.join(kn[3][8 + len(name.split()) - 1:]).split('.')[0]
     weapon = ' '.join(kn[1][6:])[:-1]
     char_adj = kn[0][-5]
+    if char_adj not in knight_adjs:
+        print("Corrupt input file.")
+        return False    
     weapon_adj = kn[1][5]
+    if weapon_adj not in weapon_adjs:
+        print("Corrupt input file.")
+        return False
     activity = list_to_string(kn[2][7 + len(name.split()):])
+    if activity not in knight_activities:
+        print("Corrupt input file.")
+        return False
     castle_adj = kn[3][4 + len(name.split()) -1]
+    if castle_adj not in castle_adjs:
+        print("Corrupt input file.")
+        return False
 
     knights[name] = {'age': age, 'weapon' : weapon, 'castle' : castle, 
                      'adjs' : [weapon_adj, castle_adj, char_adj, activity]}
@@ -468,9 +479,11 @@ def retrieve_knights(kns):
 # Takes a file and returns all the knight entries from that file (which by default is 'dead_knights.txt')
 # Also truncates the given file
 def ressurect_knights(file):
-    retrieve_knights(get_knights(read_knights(file)))
-    with open(file, 'w') as f:
-        f.truncate()
+    if retrieve_knights(get_knights(read_knights(file))):
+        with open(file, 'w') as f:
+            f.truncate()
+    else:
+        return
 
 
 #################################################### Main Menu #################################################
